sweet: fingerprint: apply a boost after fp is authenticated
Connect to the aidl power HAL extension to apply a 2ms LAUNCH boost
after fingerprint is authenticated successfully. This should speed
up the unlocking process and avoid jank in the ripple animation.
IPowerExt is used instead of the generic android.hardware.power.IPower
so that we are free to use any custom power hints that are not
defined in the generic aidl Boost type.

Some code was adapted from
hardware/google/graphics/common/libhwc2.1/libdevice/

Test: boost hint is supported and libperfmgr applies the hint

Signed-off-by: Chenyang Zhong <zhongcy95@gmail.com>

Original commit:
https://github.com/PixelExperience-Devices/device_xiaomi_sdm845-common/commit/102b2608a612fc853f7e1ed46fb7e0d31aa494bb

Signed-off-by: someone5678 <nemui3353@gmail.com>